test: added tests for AddNote and EditNote
