+++ app/src/androidTest/java/addNote/AddNoteTest.kt
@@ -1,0 +1,45 @@
+package addNote
+
+import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.onNodeWithTag
+import com.github.onlynotesswent.model.note.NoteRepository
+import com.github.onlynotesswent.model.note.NoteViewModel
+import com.github.onlynotesswent.model.users.UserRepository
+import com.github.onlynotesswent.model.users.UserViewModel
+import com.github.onlynotesswent.ui.navigation.NavigationActions
+import com.github.onlynotesswent.ui.navigation.Screen
+import com.github.onlynotesswent.ui.screen.AddNote
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.mockito.Mockito.mock
+import org.mockito.Mockito.`when`
+
+class AddNoteTest {
+  private lateinit var userRepository: UserRepository
+  private lateinit var userViewModel: UserViewModel
+  private lateinit var navigationActions: NavigationActions
+  private lateinit var noteViewModel: NoteViewModel
+  private lateinit var noteRepository: NoteRepository
+  @get:Rule val composeTestRule = createComposeRule()
+
+  @Before
+  fun setUp() {
+    // Mock is a way to create a fake object that can be used in place of a real object
+    userRepository = mock(UserRepository::class.java)
+    navigationActions = mock(NavigationActions::class.java)
+    userViewModel = UserViewModel(userRepository)
+    noteRepository = mock(NoteRepository::class.java)
+    noteViewModel = NoteViewModel(noteRepository)
+
+    // Mock the current route to be the user create screen
+    `when`(navigationActions.currentRoute()).thenReturn(Screen.ADD_NOTE)
+    composeTestRule.setContent { AddNote(navigationActions, noteViewModel) }
+  }
+
+  @Test
+  fun displayBaseComponents() {
+    composeTestRule.onNodeWithTag("AddNote text").assertExists()
+    composeTestRule.onNodeWithTag("GoBack button").assertExists()
+  }
+}
