+++ app/src/androidTest/java/editNote/EditNoteTest.kt
@@ -1,0 +1,45 @@
+package editNote
+
+import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.onNodeWithTag
+import com.github.onlynotesswent.model.note.NoteRepository
+import com.github.onlynotesswent.model.note.NoteViewModel
+import com.github.onlynotesswent.model.users.UserRepository
+import com.github.onlynotesswent.model.users.UserViewModel
+import com.github.onlynotesswent.ui.navigation.NavigationActions
+import com.github.onlynotesswent.ui.navigation.Screen
+import com.github.onlynotesswent.ui.screen.EditNote
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.mockito.Mockito.mock
+import org.mockito.Mockito.`when`
+
+class EditNoteTest {
+  private lateinit var userRepository: UserRepository
+  private lateinit var userViewModel: UserViewModel
+  private lateinit var navigationActions: NavigationActions
+  private lateinit var noteViewModel: NoteViewModel
+  private lateinit var noteRepository: NoteRepository
+  @get:Rule val composeTestRule = createComposeRule()
+
+  @Before
+  fun setUp() {
+    // Mock is a way to create a fake object that can be used in place of a real object
+    userRepository = mock(UserRepository::class.java)
+    navigationActions = mock(NavigationActions::class.java)
+    userViewModel = UserViewModel(userRepository)
+    noteRepository = mock(NoteRepository::class.java)
+    noteViewModel = NoteViewModel(noteRepository)
+
+    // Mock the current route to be the user create screen
+    `when`(navigationActions.currentRoute()).thenReturn(Screen.EDIT_NOTE)
+    composeTestRule.setContent { EditNote(navigationActions, noteViewModel) }
+  }
+
+  @Test
+  fun displayBaseComponents() {
+    composeTestRule.onNodeWithTag("EditNote text").assertExists()
+    composeTestRule.onNodeWithTag("GoBack button").assertExists()
+  }
+}
