+++ app/src/main/java/com/github/onlynotesswent/ui/screen/AddNote.kt
@@ -4,6 +4,8 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
 import com.github.onlynotesswent.model.note.NoteViewModel
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Screen
@@ -11,7 +13,11 @@ import com.github.onlynotesswent.ui.navigation.Screen
 @Composable
 fun AddNote(navigationActions: NavigationActions, noteViewModel: NoteViewModel) {
   Column {
-    Text("Add Note Screen")
-    Button(onClick = { navigationActions.navigateTo(Screen.AUTH) }) { Text("Go to Auth") }
+    Text("Add Note Screen", modifier = Modifier.testTag("AddNote text"))
+    Button(
+        onClick = { navigationActions.navigateTo(Screen.AUTH) },
+        modifier = Modifier.testTag("GoBack button")) {
+          Text("Go to Auth")
+        }
   }
 }
